Associate payment_methods and stores.
This will allow a spree installation to specify which payment methods
are used in which stores.

If a store has no payment methods associated with it, all available
payment methods are considered valid.  If a store has at least one
payment associated with it, then only associated available payment
methods are valid.

Fixes #47